% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amrfinder.R
\name{amrfinder}
\alias{amrfinder}
\title{Title}
\usage{
amrfinder(file_strains, organism, path_amrfinder, element, path_database)
}
\arguments{
\item{element}{}

\item{path_database}{}
}
\description{
Title
}
