% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_closest.R
\name{extract_closest}
\alias{extract_closest}
\title{Title}
\usage{
extract_closest(
  file_strains,
  file_querry,
  lengthconf = 95,
  identconf = 95,
  offset = 0,
  docker_image = "staphb/blast:2.15.0",
  thread = 4
)
}
\arguments{
\item{file_querry}{}
}
\description{
Title
}
