% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amrfinder.R
\name{amrfinder}
\alias{amrfinder}
\title{Title}
\usage{
amrfinder(file_strains, organism, path_amrfinder, element, path_database)
}
\arguments{
\item{file_strains}{}

\item{path_amrfinder}{the path where amrfinder is located}
}
\description{
Title
}
