% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amrfinder.R
\name{amrfinder}
\alias{amrfinder}
\title{Title}
\usage{
amrfinder(
  file_strains,
  organism,
  element,
  docker_image = "staphb/ncbi-amrfinderplus:3.12.8-2024-05-02.2",
  thread = 4
)
}
\arguments{
\item{thread}{}
}
\description{
Title
}
